requirements.txt
gunicorn
python-telegram-bot
python-telegram-bot-api
SQLAlchemy~=2.0.29
flask~=3.0.3
httpx~=0.27
pygramtic==0.2.0